test: Consolidate committable+extendable tests into single file
Merge comprehensive and edge case tests into main test file and remove
separate test scripts for better organization and maintainability.
